D general fixes (#272)
* Shas won't work

* Remove print statements

* Add TPC-H for CMake scope

* More

* Try fix

* another string issue

* compile fix

* one more

* one more

* Quick fix to check if we have issues with outputs.

* Fix query 15

* Add test to reproduce issue #269

* Add support for BitVectors.

 Fixes #269

* Add test case for #268

* Fix #268

* Create EvalFallback for SparkVeMapper

So that Wosin can use it in proj+aggregate

* Add test case for issue.

* Almost ready version of sum * smth.

* Add output support for Booleans.

* Add CASE support.

 Fixes #270

* Add CASE support.

 Fixes #270

* Support case when without else

* Simple code cleanups.

* Quick fix for #271 by skipping it.

* Undo a couple of things

* Use a hole-value

* Lift out evaluation errors so we can more precisely nail down any issues

Change fallback to be pased implicitly.
