--- aveo4j-repo/com/nec/aveo4j/0.0.1/aveo4j-0.0.1.jar.sha1
@@ -1,1 +1,0 @@
-d9b698d488813304917d7dc401e3ab47835b7585
