--- aveo4j-repo/com/nec/aveo4j/0.0.1/aveo4j-0.0.1.pom.sha1
@@ -1,1 +1,0 @@
-7300b21e04dfea6ad89c5f70b0c52886b5fc829a
