+++ src/test/scala/com/nec/spark/planning/SparkWordCountSpec.scala
@@ -39,7 +39,6 @@ final class SparkWordCountSpec extends AnyFreeSpec with BeforeAndAfter with Spar
           "SELECT word, count(word) AS count FROM words GROUP by word HAVING count > 1 ORDER by count DESC LIMIT 10"
         )
         .as[(String, BigInt)]
-    println(wordCountQuery.queryExecution.executedPlan)
     assert(wordCountQuery.collect().toList.toMap == Map("some" -> 2, "is" -> 2))
   }
